add matrix_gamma in new file
